Refactor identity management classes to their own domain.
