--- API/SummerBornInfo.Infrastructure/Persistence/ApplicationRole.cs
@@ -1,5 +1,0 @@
-namespace SummerBornInfo.Infrastructure.Persistence;
-
-public sealed class ApplicationRole : IdentityRole<Guid>
-{
-}
