--- API/SummerBornInfo.Infrastructure/Persistence/ApplicationUser.cs
@@ -1,5 +1,0 @@
-namespace SummerBornInfo.Infrastructure.Persistence;
-
-public sealed class ApplicationUser : IdentityUser<Guid>
-{
-}
