+++ API/SummerBornInfo.Infrastructure/Persistence/IdentityManagement/ApplicationRole.cs
@@ -1,0 +1,5 @@
+namespace SummerBornInfo.Infrastructure.Persistence.IdentityManagement;
+
+public sealed class ApplicationRole : IdentityRole<Guid>
+{
+}
