+++ API/SummerBornInfo.Infrastructure/Persistence/IdentityManagement/ApplicationUser.cs
@@ -1,0 +1,5 @@
+namespace SummerBornInfo.Infrastructure.Persistence.IdentityManagement;
+
+public sealed class ApplicationUser : IdentityUser<Guid>
+{
+}
